The conversion operations from uint32 to uint16 for lightmap texels were moved from d_lists into the renderer of the OXR (and the experimental PCXR) version - this is because while all lightmaps are copied to d_lists, not all lightmaps will be sent to GPU for displaying unless they were explicitly marked as updated.

+++ id1/d_lists.h
@@ -218,7 +218,7 @@ struct dlists_t
 	std::vector<unsigned char> colored_indices8;
 	std::vector<uint16_t> colored_indices16;
 	std::vector<uint32_t> colored_indices32;
-	std::vector<uint16_t> lightmap_texels;
+	std::vector<uint32_t> lightmap_texels;
 };
 
 extern dlists_t d_lists;
